requirements.txt
alembic==1.13.1
asyncpg==0.29.0
email_validator==2.1.1
fastapi==0.110.0
httptools==0.6.1
httpx==0.27.0
itsdangerous==2.1.2
Jinja2==3.1.3
orjson==3.10.0
passlib==1.7.4
pipdeptree==2.16.2
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
PyYAML==6.0.1
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
fastapi-cache2==0.2.1
celery==5.4.0
vine==5.1.0
flower==2.0.1
tornado==6.4
pillow==10.3.0
